requirements.txt
asgiref==3.4.1
Django==3.2.8
django-cors-headers==3.10.0
django-environ==0.8.1
django-rest-framework-nested==0.0.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Pillow==8.4.0
psycopg2==2.9.1
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
